Move injector patches to a separate file
